+++ src/App.css
@@ -3,7 +3,7 @@
   padding: 0;
   box-sizing: border-box;
 
-  /* border: 1px solid red; */
+  /* outline: 1px solid rgba(255, 0, 0, 0.2); */
 }
 
 html {
