+++ src/component/Header.css
@@ -24,7 +24,7 @@
     border-top: 1px solid white;
 
     background: #2A4D5F;
-    color: #F7F5F2;
+    color: white;
   }
 
   .subNavList {
@@ -90,14 +90,14 @@
     padding: 0.7vw 1.5vw;
     border-radius: 1vw;
     cursor: pointer;
-    border: 1px solid #F7F5F2;
-    color: #F7F5F2;
+    border: 1px solid white;
+    color: white;
     font-size: 1em;
     font-weight: 500;
   }
 
   .rightMenu .reserve-small:hover {
-    background: #F7F5F2;
+    background: white;
     color: #2A4D5F;
   }
 
@@ -131,7 +131,7 @@
     transition: transform 0.5s ease-in-out;
     
     list-style-type: none;
-    background: #F7F5F2;
+    background: white;
   }
 
   .burger-logIn {
@@ -143,7 +143,7 @@
   }
 
   .burger-logIn:hover {
-    color: #F7F5F2;
+    color: white;
     background: #C7925D;
   }
 
@@ -163,7 +163,7 @@
 /* 사이즈 공통 */
 .nav {
   background-color: #2A4D5F;
-  color: #F7F5F2;
+  color: white;
   
   width: 100%;
   height: 70px;
