+++ src/component/ReservationComponent/RsvnNextBtn/RsvnNextBtn.css
@@ -1,4 +1,4 @@
-.rsvnNextBtn {
+.rsvnNextBtn, .rsvnPayNextBtn {
   display: flex;
   justify-content: space-between;
   width: 100%;
