+++ src/component/ReservationComponent/RsvnNowInfo/NowInfoDefault.css
@@ -1,17 +1,3 @@
-/* pc 사이즈 : 네비게이션 바 */
-@media screen and (min-width:768px) {
-  :root {
-    font-size: 16px;
-  }
-}
-
-/* 모바일 사이즈 : 햄버거 버튼 */
-@media screen and (max-width:767px) {
-  :root {
-    font-size: 10px;
-  }
-}
-
 .rsvnInfoNow {
   height: 100%;
 }
