+++ src/component/ReservationComponent/RsvnPay/RsvnPay.css
@@ -1,0 +1,5 @@
+.creditTab {
+  padding: 10px;
+  border: 1px solid black;
+  cursor: pointer;
+}
