+++ src/pages/ReservationPage/RsvnRoomPage.css
@@ -28,5 +28,5 @@
 }
 
 .roomLayout-right {
-  flex: 1;
+  flex: 1.1;
 }
